+++ vlib/builtin/string_test.v
@@ -621,15 +621,15 @@ fn test_capitalize() {
 }
 
 fn test_title() {
-	mut s := 'hello world'
+	mut s := 'hello world!'
 	assert !s.is_title()
-	assert s.title() == 'Hello World'
+	assert s.title() == 'Hello World!'
 	s = 'HELLO WORLD'
 	assert !s.is_title()
-	assert s.title() == 'HELLO WORLD'
-	s = 'Hello World'
+	assert s.title() == 'HELLO WORLD!'
+	s = 'Hello World!'
 	assert s.is_title()
-	assert s.title() == 'Hello World'
+	assert s.title() == 'Hello World!'
 }
 
 fn test_for_loop() {